:art: :: upload 로직 분리
