+++ src/main/java/com/bamboo/bamboodgsw/domain/post/presentation/PostController.java
@@ -7,10 +7,7 @@ import com.bamboo.bamboodgsw.domain.post.presentation.ro.PostListRo;
 import com.bamboo.bamboodgsw.domain.post.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 @RestController
 @RequiredArgsConstructor
@@ -41,18 +38,4 @@ public class PostController {
         return postService.getPostById(postId);
     }
 
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/attachment")
-    public Long uploadAttachment(
-            @RequestParam MultipartFile request
-    ) {
-        return postService.uploadAttachment(request);
-    }
-
-    @GetMapping(value = "/attachment{attachment-id}", produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
-    public ResponseEntity<byte[]> getAttachment(
-            @PathVariable("attachment-id") Long attachmentId
-    ) {
-        return postService.getAttachment(attachmentId);
-    }
 }
